requirements.txt
streamlit
openai-whisper
spacy
ffmpeg-python
numpy
pandas
torch
transformers
pydub
pyaudio
scipy




